added count balancer

+++ Final Project 01-02-23/Downloader/src/services/file.services.ts
@@ -1,25 +1,37 @@
 import FileEntity from '../database/entities/file.entity'
 import { FileRepository } from '../database/repository/file.repository'
+import FileAccountService from './file-account.service'
 
 export default class FileService {
   private fileRepository: FileRepository
+  private fileAccountService: FileAccountService
+
   constructor () {
     this.fileRepository = new FileRepository()
+    this.fileAccountService = new FileAccountService()
   }
 
-  async createFileById (newFile: any) {
-    const contentLinks = newFile.driveFile.map((entry:any) => {
-      return entry.contentLink
-    })
-    const formatFile = new FileEntity()
-    formatFile.name = newFile.name
-    formatFile.status = newFile.status
-    formatFile.size = newFile.size
-    formatFile.contentLinks = contentLinks
-    formatFile.uploaderId = newFile.id.toString()
+  async updateFileFromUploader (File: any) {
+    const fileFromDB: FileEntity | undefined = await this.getFileByUploaderId(File.id)
+    const fileToUpdate: FileEntity = fileFromDB || new FileEntity()
 
-    const newEntry = await this.fileRepository.createFile(formatFile)
-    console.log(newEntry)
+    fileToUpdate.name = File.name
+    fileToUpdate.uploaderId = File.id.toString()
+    fileToUpdate.size = File.size
+    fileToUpdate.downloadsToday = fileToUpdate.downloadsToday | 0
+    fileToUpdate.downloadsTotal = fileToUpdate.downloadsTotal | 0
+
+    return await this.fileRepository.updateFile(fileToUpdate)
+  }
+
+  async updateFileFromDownloader (File: any) {
+    const fileFromDB: FileEntity | undefined = await this.getFileById(File.id)
+    const fileToUpdate: FileEntity = fileFromDB || new FileEntity()
+
+    fileToUpdate.downloadsToday = File.downloadsToday | 0
+    fileToUpdate.downloadsTotal = File.downloadsTotal | 0
+
+    return await this.fileRepository.updateFile(fileToUpdate)
   }
 
   async getAllFiles () {
@@ -36,7 +48,15 @@ export default class FileService {
 
   async deleteFile (id:string) {
     const foundFile = await this.getFileByUploaderId(id)
-    const deletedFile = await this.fileRepository.deleteFile(foundFile.id)
-    console.log(deletedFile)
+
+    if (foundFile) {
+      await this.fileRepository.deleteFile(foundFile.id)
+
+      const fileRelationsWithAccounts = await this.fileAccountService.getRelationstByFileId(foundFile.uploaderId)
+
+      fileRelationsWithAccounts.forEach((relation) => {
+        this.fileAccountService.deleteFileAccount(relation.id)
+      })
+    }
   }
 }
